Изменение таксономии, чтобы пройти PR автотесты
